Added Shared EXP Level Preservation
No matter what the growth rate of the class, they will all be the same
level and percentage to the next level. THe amount of EXP taken to level
is not changed, changing levels simply preserves teh current percentage
of EXP relative to the amount the new class needs to level.